+++ packages/plugin-ext/src/plugin/debug/plugin-debug-adapter-session.ts
@@ -14,9 +14,9 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import { DebugAdapterSessionImpl } from '@theia/debug/lib/node/debug-adapter-session';
+import { DebugAdapterSessionImpl } from '@theia/debug/lib/common/debug-adapter-session';
 import * as theia from '@theia/plugin';
-import { DebugAdapter } from '@theia/debug/lib/node/debug-model';
+import { DebugAdapter } from '@theia/debug/lib/common/debug-model';
 import { DebugChannel } from '@theia/debug/lib/common/debug-service';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
